+++ cpp/libjoynr/include/joynr/CapabilitiesRegistrar.h
@@ -136,7 +136,7 @@ public:
         messageRouter->removeNextHop(participantId, callbackFct);
         future->wait();
 
-        if (!future->getStatus().success()) {
+        if (!future->isOk()) {
             JOYNR_LOG_ERROR(logger,
                             "Unable to remove next hop (participant ID: {}) from message router.",
                             participantId);